Move the coords comp out of the core map
